api/requirements.txt
alembic==1.10.3
anyio==3.7.1
asgiref==3.6.0
asyncpg==0.28.0
cachetools==4.2.4
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
dnspython==2.3.0
email-validator==2.0.0
fastapi==0.103.2
google-auth==2.17.3
# greenlet==2.0.2
h11==0.14.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.4
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.21.1
phonenumbers==8.13.9
bcrypt==4.0.1
redis==4.5.5
httpx==0.24.1
requests==2.31.0
jinja2==3.1.2
python-multipart==0.0.6
